// src/api/endpoints/api.schemas.ts
export interface RenovationPerms {
  renovationId?: string;
  /** @nullable */
  userId?: string | null;
  canEdit?: boolean;
}